新增接口:点击去结算,结算购物车

+++ entity/basket/req/addProdToBasketReq.go
@@ -2,7 +2,7 @@ package req
 
 type AddProdToBasketReq struct {
     BasketId uint64 `json:"basket_id;"`
-    ShopId   int    `json:"shopId"`   //  店铺id
+    ShopId   uint64    `json:"shopId"`   //  店铺id
     UserId   uint64 `json:"userId"`   // 用户id
     SkuId    uint64 `json:"skuId"`    // 单品ID
     ProdNums int    `json:"prodNums"` // 商品数量
